frontend-common:  Only make a FIFO if we'll use it
Also, add the missing <sys/stat.h> include.

+++ graphics/frontend-common/frend/raster_updating.hh
@@ -7,10 +7,12 @@
 #define FREND_RASTERUPDATING_HH
 
 // POSIX
+#include <sys/stat.h>
 #include <unistd.h>
 
 // frontend-common
 #include "frend/displayfs.hh"
+#include "frend/update_fifo.hh"
 
 
 namespace frend
@@ -26,12 +28,18 @@ class raster_updating
 	public:
 		raster_updating()
 		{
-			mkfifo( update_fifo, 0666 );
+			if ( CONFIG_UPDATES_VIA_FIFO )
+			{
+				mkfifo( update_fifo, 0666 );
+			}
 		}
 		
 		~raster_updating()
 		{
-			unlink( update_fifo );
+			if ( CONFIG_UPDATES_VIA_FIFO )
+			{
+				unlink( update_fifo );
+			}
 		}
 };
